// src/core/post/dto/post-find-params.dto.ts
import { LangType } from 'src/common/enum/lang.enum';
import { PaginationType } from 'src/common/pipe/page-validation.pipe';

export class findAllPostParamsDto extends PaginationType {
  sort: string;
  by: 'DESC' | 'ASC';
  where: string;
  category: string;
  lang: LangType;
  userId?: number;
  getAll?: boolean;
}

export type findOnePostParamsDto = {
  id: string;
  userId?: number;
};